+++ NEWS.rst
@@ -150,6 +150,7 @@ Changes
   expected frame format.
 * ``ZstdCompressor`` now has a ``frame_progression()`` method to return
   information about the current compression operation.
+* Error messages in CFFI no longer have ``b''`` literals.
 
 0.8.2 (released 2018-02-22)
 ---------------------------